+++ .storybook/preview.tsx
@@ -1,0 +1,31 @@
+import type { Preview } from '@storybook/react';
+import React from 'react';
+import { MemoryRouter } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import '../src/index.css';
+
+const queryClient = new QueryClient();
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    Story => {
+      return (
+        <MemoryRouter>
+          <QueryClientProvider client={queryClient}>
+            <Story />
+          </QueryClientProvider>
+        </MemoryRouter>
+      );
+    },
+  ],
+};
+
+export default preview;
