+++ src/mocks/browser.ts
@@ -5,7 +5,6 @@ import { loginHandlers } from '@/mocks/handler/loginHandlers';
 import { tokenHandlers } from '@/mocks/handler/tokenHandlers';
 import { totalHandlers } from '@/mocks/handler/totalHandlers';
 import { transactionHandlers } from '@/mocks/handler/transactionHandlers';
-import { tokenVerifyHandlers } from '@/mocks/handler/tokenVerifyHandler';
 import { categoryHandlers } from '@/mocks/handler/categoryHandlers';
 import { chartHandlers } from '@/mocks/handler/chartHandlers';
 import { categoryLogsHandler } from './handler/categoryLogsHandler';
@@ -15,7 +14,6 @@ export const worker = setupWorker(
   ...tokenHandlers,
   ...authHandlers,
   ...emailHandlers,
-  ...tokenVerifyHandlers,
   ...totalHandlers,
   ...transactionHandlers,
   ...categoryHandlers,
